DOX/Orion-Agent/sr-analyzer-master/semantic-resolution/requirements.txt
# SR Feedback System - Clean Environment Requirements
# This is the working configuration with FULL sklearn support

# Web Framework
flask==3.1.2
werkzeug==3.1.3

# Data Processing
pandas>=2.0.0
numpy<2.0.0  # Required for torch/sentence-transformers compatibility

# Machine Learning (for semantic search and TF-IDF vectorization)
scikit-learn>=1.3.0

# PyTorch (pinned for sentence-transformers compatibility)
torch==2.2.2

# Excel File Handling
openpyxl==3.1.5  # For .xlsx files
xlrd==2.0.2      # For .xls files

# Sentence Transformers for RAG embeddings (pinned for stability)
sentence-transformers==2.5.1

# ChromaDB for vector store
chromadb>=0.4.0

# Date/Holiday calculations
holidays>=0.40

# Windows COM automation (for Outlook integration)
pywin32>=306

# HTTP requests for LLM API calls
requests>=2.31.0

# Progress bars for batch processing
tqdm>=4.65.0

# LangChain for LLM integration
langchain>=0.1.0
langchain-core>=0.1.0

# PostgreSQL driver for ActivityFinder
psycopg2-binary>=2.9.0

# Azure AD Authentication
msal>=1.24.0
python-dotenv>=1.0.0

# Microsoft Graph API for Outlook email integration (Linux)
O365>=2.0.0